refactoring(src): изменена структура приложения
