+++ CircuitConstructor/CircuitConstructor/ClientApp/src/components/AddShapeButton.jsx
@@ -1,5 +1,4 @@
 import React from "react";
-import "./styles/canvasStyles.css";
 
 /**
  * Кнопка добавления элемента на Canvas.
