+++ CircuitConstructor/CircuitConstructor/ClientApp/src/components/TopMenu.jsx
@@ -1,22 +1,21 @@
 import React from 'react';
-import {ReactComponent as SaveFileIcon} from "./svgElements/interfaceElements/SaveFileIcon.svg";
-import {ReactComponent as UploadFileIcon} from "./svgElements/interfaceElements/UploadFileIcon.svg";
-import "./styles/mainContentStyles.css";
+import {ReactComponent as SaveFileIcon} from "../assets/interfaceElements/SaveFileIcon.svg";
+import {ReactComponent as UploadFileIcon} from "../assets/interfaceElements/UploadFileIcon.svg";
 
 /**
  * Верхняя панель с кнопками сохранения и загрузки.
  */
 const TopMenu = () => {
-        return (
-            <div className="top-panel">
-                <button className="top-panel-button">
-                    <SaveFileIcon />
-                </button>
-                <button className="top-panel-button">
-                    <UploadFileIcon />
-                </button>
-            </div>
-        );
+    return (
+        <div className="top-panel">
+            <button className="top-panel-button">
+                <SaveFileIcon />
+            </button>
+            <button className="top-panel-button">
+                <UploadFileIcon />
+            </button>
+        </div>
+    );
 }
 
 export default TopMenu;
