+++ CircuitConstructor/CircuitConstructor/ClientApp/src/hooks/useComponentVisible.jsx
