+++ CircuitConstructor/CircuitConstructor/ClientApp/src/styles/mainContentStyles.css
