+++ CircuitConstructor/CircuitConstructor/ClientApp/src/styles/sidebarStyles.css
